requirements.txt
absl-py==2.1.0
accelerate==0.28.0
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
annotated-types==0.6.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
basicsr==1.4.2
cachetools==5.4.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
cmake==3.25.0
contourpy==1.0.5
crcmod==1.7
cryptography==42.0.8
cycler==0.12.1
datasets==2.18.0
deepspeed==0.14.0
diffusers==0.27.2
dill==0.3.8
dnspython==2.6.1
einops==0.8.0
email_validator==2.1.1
exceptiongroup==1.2.1
facexlib==0.3.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.9.0
filterpy==1.4.5
fonttools==4.37.4
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.5.4
gfpgan==1.3.8
google-auth==2.33.0
google-auth-oauthlib==1.0.0
gpytorch==1.12
grpcio==1.49.1
h11==0.14.0
h5py==3.11.0
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.4
imageio==2.34.2
importlib-metadata==6.3.0
importlib_resources==6.4.0
jaxtyping==0.2.19
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.4
lime==0.2.0.1
linear-operator==0.5.2
lit==15.0.7
llvmlite==0.41.1
lmdb==1.3.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.7.5
mdurl==0.1.2
modelscope==1.15.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
netcal==1.3.5
networkx==3.1
ninja==1.11.1.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
orjson==3.10.4
oss2==2.18.5
packaging==21.3
pandas==2.0.3
pillow==10.4.0
protobuf==5.27.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.18.0
pynvml==11.5.0
pyparsing==3.1.2
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
realesrgan==0.3.0
regex==2023.12.25
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.10.1
seaborn==0.13.0
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
sympy==1.12
tb-nightly==2.14.0a20230808
tensorboard==2.12.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2023.7.10
tikzplotlib==0.9.8
tokenizers==0.19.1
torch==2.4.0
torchaudio==0.13.0+cu117
torchmetrics==0.11.4
torchvision==0.19.0
tqdm==4.66.5
transformers==4.41.2
triton==3.0.0
typeguard==2.13.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==1.26.13
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
xxhash==3.4.1
yapf==0.32.0
yarl==1.9.4
zipp==3.19.2
